requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
pyyaml>=6.0
scikit-learn>=1.0.0
jupyter>=1.0.0
pytest>=7.0.0
